Refactor Map namespace in Map.h and Map.cpp
